+++ src/promoted/code1_0019.c
@@ -38,7 +38,7 @@ extern u8 *iGpffffb3b8;
 extern u16 func_001d7f10(u8 *arg0, u8 *arg1, u16 arg2, u32 arg3);
 extern s32 func_001f0a50(u8 *arg0);
 extern s32 func_001f11e0(s64 arg0);
-extern s32 func_0023d8e0(u8 *arg0, s32 arg1);
+extern s32 func_0023d8e0(u8 *arg0, u16 arg1);
 extern u8 *iGpffffb3c0;
 extern u8 *iGpffffb3cc;
 extern s64 func_00192560(void);
